a problem soled by  Cpp power
on a gloomy and pointless day after emmbaressing , useless night i decide to solve a problem i recently was seen but i couldn't solve it by myself. so i deciede to write an algoritm for it and code it to cpp.
i used simple algorithm with 3 internal for. i did it at the simplest and fast i can.
 another clu was use a recersive function.